goauth2: adding a jwt package to support server to server applications.
The jwt package allows the user to define a jwt.Token containing the
required and optional fields for constructing the serialized and encoded
JWT to be used for requesting an access_token.  Integrating this package
into oauth/oauth.go will require some redesign of the parent package.
These changes have been postponed for a later CL.

The descriptions of the standards this change implements can be found on
the following web page.

https://developers.google.com/accounts/docs/OAuth2ServiceAccount

This addition to the library will allow for example a GAE instance to
make requests to service like BigQuery.

R=adg
CC=golang-dev
http://codereview.appspot.com/6496045

Committer: Andrew Gerrand <adg@golang.org>